+++ src/tool/offline_server/server.rs
@@ -103,6 +103,13 @@ pub async fn offline_rpc_state(
         state_manager.chain_config().eth_chain_id,
         message_pool.subscriber(),
     ));
+    services.spawn({
+        let eth_event_handler = eth_event_handler.clone();
+        async move {
+            eth_event_handler.run_filter_gc().await;
+            Ok(())
+        }
+    });
     Ok((
         RPCState {
             state_manager,